backend/requirements.txt
fastapi>=0.115,<1.0
networkx>=3.4,<4.0
pydantic>=2.8,<3.0
pyyaml>=6.0,<7.0
uvicorn[standard]>=0.30,<1.0
pytest>=8.0,<9.0
